+++ Gruntfile.js
@@ -92,6 +92,10 @@
                     "src/idd/idd.onscreennavigation.js",
                     "src/idd/idd.bingmapsplot.js",
                     "src/idd/idd.ko.js",
+                    "src/idd/idd.ko.markers.js",
+                    "src/idd/idd.ko.polyline.js",
+                    "src/idd/idd.ko.domplot.js",
+                    "src/idd/idd.ko.area.js",
                     "src/idd/wrapper_footer_knockout.txt"
                 ],
                 dest: 'dist/idd_knockout.js',